Lab 5 second try
Asked Mike: In our osMutexAcquire(), we just check if the mutex is available or not. If not, then it sets puts that thread in the waiting queue, and then sets its status to BLOCKED. However, I realized that that does not actually stop the rest of the thread from running (it sets the status to BLOCKED, but then the following lines run anyway.) Our osMutexAcquire() returns a bool value, so would it be correct to use an IF statement to check the mutex before running anything else (seems like it might be weird to me, and when I tested it with a "false" output from trying to obtain the mutex, the code seems to hard fault). Is there a better solution to this? I've attached screenshots of what I mean.

Also, preemptive switching seems to break on thread 3